+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.0.24
+
+* Added third parameter to substitute(), allowing original URL to be kept even where the substituting route has its own URL
+* Allowing postponeDefault() to take a callback
+
 ## 0.0.23
 
 * Fixed route timestamp